fix: removing used regions

+++ courier-backend/src/main/java/pl/polsl/tab/kurier/service/RegionService.java
@@ -3,7 +3,10 @@ package pl.polsl.tab.kurier.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.polsl.tab.kurier.dto.RegionDTO;
+import pl.polsl.tab.kurier.exception.ResourceBusyException;
 import pl.polsl.tab.kurier.model.Region;
+import pl.polsl.tab.kurier.repository.AddressRepository;
+import pl.polsl.tab.kurier.repository.ParcelRepository;
 import pl.polsl.tab.kurier.repository.RegionRepository;
 
 import java.util.List;
@@ -14,6 +17,12 @@ public class RegionService {
 
     @Autowired
     private RegionRepository regionRepository;
+
+    @Autowired
+    private AddressRepository addressRepository;
+
+    @Autowired
+    private ParcelRepository parcelRepository;
 
     public List<RegionDTO> getAllRegions() {
         return regionRepository.findAll().stream()
@@ -47,6 +56,13 @@ public class RegionService {
     }
 
     public void deleteRegion(Integer id) {
+        if (addressRepository.countByRegionRegionId(id) > 0) {
+            throw new ResourceBusyException("Cannot delete region: it is associated with existing addresses.");
+        }
+        if (parcelRepository.countByCurrentRegionRegionId(id) > 0 || parcelRepository.countByNextRegionRegionId(id) > 0) {
+            throw new ResourceBusyException("Cannot delete region: it is associated with active parcels.");
+        }
+        
         regionRepository.deleteById(id);
     }
 }
